split modular and namespaced. import into index
